i'm wondering if I should just merge this whole package back into x509
